requirements.txt
streamlit
pyPDF2
google.generativeAI
python-dotenv
fpdf